fix: code corrections
